Harmonized core advection routine to use the same code over each 1-d row.
Changed ns_aloft to n_gs_aloft to be specifically tied to grainsizes and not
species.

Changed Collapse_GS to Prune_GS which just tags slices of the concentration
array as flushed out instead of reordering slices.  This is because sorting
by fall velocity was problematic.

Removed (or at least commented out) Tephra subrouting Sort_Tephra_Size since
this was only needed to sorting by fall velocity, which was problematic when
having aggregate classes.

Added initialization of vf_sp to 0.0 in fall velocity calculations.