Changed write_NETCDF.f90 so that it is a module (Ash3d_Netcdf).  This will allow the subroutines in this file to be used in a post-processing program. Added variables and dimensions to account for all output products including Airport/POI variables and vertical profiless.  Added subroutine NC_check_status to handle all errors and clean up the code.  Added subroutine NC_Read_Output_Products which is used in the post-processing program.  Added check if the netcdf library supports v4 and then uses compression.
Moved Adjust_DT to Mesointerpolator and added an optional logical input parameter which when TRUE will calculate the appropriate dt using velocities at the the next time step of the windfiles.  Using NWP timestep dt values and interpolating dramatically reduces the number of times Adjust_DT needs to be called.  Time steps, when only dependent on NWP data, will still be shorter than the dt's calculated at each step.  Using this feature is invoked if the preprocessor flag FAST_DT is set.

Added a new subroutine (get_minmax_index) to Calc_Mesh which finds the minimum and maximum extends of the ash cloud, based on CLOUDCON_GRID_THRESH = 1.0e-7. This is invoked if the preprocessor flag FAST_SUBGRID is set.  When this is set, advect_x, advect_y, and advect_z all execute on just this subgrid.  This subgrid is also used in Ouput_Vars for calculations of dbZCalculator, ConcentrationCalculator, AshloadCalculator.

Added CLOUDCON_GRID_THRESH to the ResetParams.f90 list.