Re-wrote diffusion routines with proper use of cell volumes and boundary conditions. Explicit solver supports periodicity, but not implicit. Implicit can be adjusted from FE to CN to BE by setting Imp_fac Sub-grid is set up but not fully implemented.
Reactivated vertical diffusion in top file.
Fixed bug in sigma_nz where r was off by 0.5 dz.
Changed Adjust_dt so that the diffusion constraint is only calculated once if kx,ky,kz
are fixed in time.  Also added a constant factor to guide the DT selection for
when velocities are very low.  Imp_DT_fac is the multiples of dx2/k that are allowed.
Explicit diffusion requires 0.5, but this is set to 4.0.